Refactor index.py into smaller modules
---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/grp06/cursor-boost?shareId=XXXX-XXXX-XXXX-XXXX).
