+++ snapshot_writer.py
@@ -1,0 +1,5 @@
+import os
+
+def write_snapshot(snapshot):
+    with open("snapshot.txt", "w") as f:
+        f.write(snapshot)
